requirements.txt
# Deep Learning Framework
torch==2.0.1
torchvision==0.15.2

# Data Processing
numpy==1.24.3
Pillow==10.0.0

# Visualization
matplotlib==3.7.2

# Machine Learning Utilities
scikit-learn==1.3.0

# Web Framework (for API service)
flask==2.3.2
flask-cors==4.0.0

# Image Processing
opencv-python==4.8.0.74

# Utilities
tqdm==4.65.0
